backend/requirements.txt
Django
djangorestframework
djoser
webcolors
django-cors-headers
Pillow
django-filter
python-dotenv==1.0.0
psycopg2-binary